requirements.txt
altair=5.2.0=pypi_0
attrs=23.2.0=pypi_0
blinker=1.7.0=pypi_0
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.12.12=hca03da5_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
fonttools=4.47.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.41=pypi_0
idna=3.6=pypi_0
importlib-metadata=6.11.0=pypi_0
jinja2=3.1.3=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.5=pypi_0
libffi=3.4.4=hca03da5_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
mdurl=0.1.2=pypi_0
ncurses=6.4=h313beb8_0
numpy=1.26.3=pypi_0
openssl=3.0.12=h1a28f6b_0
packaging=23.2=pypi_0
pandas=2.1.4=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py310hca03da5_0
protobuf=4.25.2=pypi_0
pyarrow=14.0.2=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
python=3.10.13=hb885b13_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2023.3.post1=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.32.1=pypi_0
requests=2.31.0=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.16.2=pypi_0
setuptools=68.2.2=py310hca03da5_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h80987f9_0
streamlit=1.29.0=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.3=pypi_0
tk=8.6.12=hb8d0fd4_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.4=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.4=pypi_0
tzlocal=5.2=pypi_0
urllib3=2.1.0=pypi_0
validators=0.22.0=pypi_0
wheel=0.41.2=py310hca03da5_0
xz=5.4.5=h80987f9_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5a0b063_0
